+++ core/process.js
@@ -51,6 +51,37 @@ Process.prototype.setup = function () {
 	this.emit('nocluster.setup');
 };
 
+// public this is available for worker processes in cluster mode only 
+Process.prototype.send = function (msg, cb) {
+	if (!this.inClusterMode) {
+		// we don't do anything in none-cluster mode
+		this.log.warn('gracenode.send() is not available in nonde-cluster mode');
+		return cb();
+	}
+	if (cluster.isMaster) {
+		return cb(new Error('gracenode.send() is available for worker processes only'));
+	}
+	var data = null;
+	try {
+		data = JSON.stringify(msg);
+	} catch (e) {
+		return cb(e);
+	}
+	// set up the listener for response
+	process.once('message', function (data) {
+		if (typeof cb === 'function') {
+			try {
+				data = JSON.parse(data);
+			} catch (e) {
+			
+			}
+			cb(null, data);
+		}
+	});
+	// send the message to master
+	process.send(data);
+};
+
 // private	
 Process.prototype.startClusterMode = function () {
 	if (cluster.isMaster) {
@@ -63,8 +94,20 @@ Process.prototype.startClusterMode = function () {
 
 // private (master only)
 Process.prototype.createWorker = function () {
+	// create worker process
 	var worker = cluster.fork();
 	this.log.info('worker spawned (pid: ' + worker.process.pid + ')');
+	// set up message listener for the worker
+	var that = this;
+	worker.on('message', function (data) {
+		try {
+			data = JSON.parse(data);
+		} catch (e) {
+			
+		}
+		that.gracenode.emit('worker.message', worker, data);
+	});
+	this.log.verbose('worker message listener set up complete: (pid:' + worker.process.pid + ') (id:' + worker.id + ')');
 	return worker;
 };
 
@@ -118,6 +161,15 @@ Process.prototype.setupWorker = function () {
 		that.log.info('worker (pid:' + process.pid + ') has disconnected');
 		that.emit('shutdown');
 		that.gracenode.exit();
+	});
+	// set up message lisetner
+	process.on('message', function (data) {
+		try {
+			data = JSON.prase(data);
+		} catch (e) {
+		
+		}
+		that.gracenode.emit('master.message', cluster.worker, data);
 	});
 };